Partial refactoring of import and export
