+++ internal/domain/import.go
@@ -27,6 +27,10 @@ package domain
 
 import "github.com/spf13/cobra"
 
+func (d *Domain) ImportInit(cmd *cobra.Command, args []string) error {
+	return d.externalInventoryProvider.ImportInit(cmd, args, d.datastore)
+}
+
 func (d *Domain) Import(cmd *cobra.Command, args []string) error {
 	return d.externalInventoryProvider.Import(cmd, args, d.datastore)
 }
